feat(components): migrate Hr to TypeScript
feature/typescript-hr
