+++ src/components/Hr/Hr.docs.mdx
@@ -1,5 +1,5 @@
 import { Status, Props, Story } from '../../../.storybook/components';
-import Hr from '../../../src/components/Hr/Hr';
+import { Hr } from './Hr';
 
 # HR
 
@@ -8,13 +8,11 @@ import Hr from '../../../src/components/Hr/Hr';
 HR provides a visual line that separates content into sections
 
 <Story id="components-hr--base" />
-
 <Props of={Hr} />
 
 ## When to use it
 
-Use it when you need to separate different sections of a certain content within a body text or a list
-of data within a card.
+Use it when you need to separate different sections of a certain content within a body text or a list of data within a card.
 
 ## Usage guidelines
