+++ src/components/Hr/Hr.spec.tsx
@@ -13,6 +13,27 @@
  * limitations under the License.
  */
 
-import Hr from './Hr';
+import React from 'react';
 
-export default Hr;
+import { create, renderToHtml, axe } from '../../util/test-utils';
+
+import { Hr } from './Hr';
+
+describe('Hr', () => {
+  /**
+   * Style tests.
+   */
+  it('should render with default styles', () => {
+    const actual = create(<Hr />);
+    expect(actual).toMatchSnapshot();
+  });
+
+  /**
+   * Accessibility tests.
+   */
+  it('should meet accessibility guidelines', async () => {
+    const wrapper = renderToHtml(<Hr />);
+    const actual = await axe(wrapper);
+    expect(actual).toHaveNoViolations();
+  });
+});
