+++ src/components/Hr/index.ts
@@ -13,25 +13,6 @@
  * limitations under the License.
  */
 
-import React from 'react';
+import { Hr } from './Hr';
 
-import Hr from './Hr';
-
-describe('Hr', () => {
-  /**
-   * Style tests.
-   */
-  it('should render with default styles', () => {
-    const actual = create(<Hr />);
-    expect(actual).toMatchSnapshot();
-  });
-
-  /**
-   * Accessibility tests.
-   */
-  it('should meet accessibility guidelines', async () => {
-    const wrapper = renderToHtml(<Hr />);
-    const actual = await axe(wrapper);
-    expect(actual).toHaveNoViolations();
-  });
-});
+export default Hr;
